Licença ao projeto e mudança na estrutura interna de pacotes.
