+++ META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
-Manifest-Version: 1.3.3
-Main-Class: br.com.gaboso.Scriptum
+Manifest-Version: 1.3.4
+Main-Class: com.github.gaboso.Scriptum
