+++ pom.xml
@@ -6,7 +6,27 @@
 
     <groupId>br.com.gaboso</groupId>
     <artifactId>scriptum</artifactId>
-    <version>1.3.3</version>
+    <version>1.3.4</version>
+
+    <organization>
+        <name>com.github.gaboso</name>
+        <url>https://github.com/Gaboso/Scriptum</url>
+    </organization>
+    <licenses>
+        <license>
+            <name>GNU General Public License, version 3</name>
+            <url>http://www.gnu.org/licenses/gpl-3.0.txt</url>
+            <distribution>repo</distribution>
+        </license>
+    </licenses>
+    <scm>
+        <connection>
+            scm:git:git://github.com/Gaboso/Scriptum.git
+        </connection>
+        <url>
+            https://github.com/Gaboso/Scriptum
+        </url>
+    </scm>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -33,7 +53,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>br.com.gaboso.Scriptum</mainClass>
+                            <mainClass>com.github.gaboso.Scriptum</mainClass>
                         </manifest>
                     </archive>
                     <descriptorRefs>
