+++ src/main/java/com/github/gaboso/module/GitModule.java
@@ -1,8 +1,8 @@
-package br.com.gaboso.module;
+package com.github.gaboso.module;
 
-import br.com.gaboso.format.Formatter;
-import br.com.gaboso.helper.CommandHelper;
-import br.com.gaboso.helper.OsHelper;
+import com.github.gaboso.format.Formatter;
+import com.github.gaboso.helper.CommandHelper;
+import com.github.gaboso.helper.OsHelper;
 import org.apache.log4j.Logger;
 
 import java.io.BufferedReader;
